Prune the validation apparatus to its property core
The PR carried ~2.7k lines of tests and benches against a ~280-line
net library change. Pruned by the criterion: keep tests that check
general properties against oracles, drop demonstrations and anything
that locks in current behavior.

Removed:
* examples/int_proxy_bench.rs — a benchmark, not an example. Its
  numbers and their analysis are recorded in the branch's commit
  messages; the harness is recoverable from history when next needed.
* tests/int_proxy_bytes.rs — the [u8]-rows recipe was a demonstration
  wearing a test costume. Its genuinely general content (collision
  resolution under compressing tokens) goes with it for now; the
  hashed-token instantiation is currently uncovered, noted as a gap
  for a slimmer future test.

Consolidated tests/{common/mod.rs, int_proxy_graphs.rs,
int_proxy_pending.rs} into one tests/int_proxy.rs (821 lines) keeping
the property tests: join vs a naive oracle, reduce vs exact per-time
counting, the randomized differential fuzz, the multi-time slow-path
oracle, the partial-order pending contract, and interleaved-unit
isolation (which pinned the Sink bug and is verified to fail against
the bug shape).

examples/scc_proxy.rs stays: it is an actual example (how to write a
backend and mount the tactic in a dataflow), and its SCC_VALIDATE
oracle is the deepest integration check we have.
